% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addins}
\alias{addins}
\alias{str_addin}
\alias{head_addin}
\alias{View_addin}
\alias{summary_addin}
\alias{dim_addin}
\alias{class_addin}
\title{Rstudio keyboard shortcuts on F-keys}
\usage{
str_addin(obj = selectobject())

head_addin(obj = selectobject())

View_addin(obj = selectobject())

summary_addin(obj = selectobject())

dim_addin(obj = selectobject())

class_addin(obj = selectobject())
}
\arguments{
\item{obj}{List containing \code{object} (some R object) \code{fullcode}
(code, objectname, expression) and \code{code} (potentially truncated version).
DEFAULT: Rstudio addin selected code from \code{\link{selectobject}}}
}
\value{
Output of the respective functions
}
\description{
Rstudio addins to examine highlighted code / object.
It can be very useful to bind these addins as outlined in \url{https://github.com/brry/rstudioberry#rstudio-addins-for-keyboard-shortcuts}
}
\examples{
# Go to Addins - browse Addins - Keyboard shortcuts - map commands as desired

}
\references{
Heavily borrowed from \url{https://github.com/digital-dharma/RStudioAddIns}
            To be added to \url{https://github.com/daattali/addinslist#readme}
}
\seealso{
\code{\link{selectobject}, \link{funSource}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2017
}
