% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdoc.R
\name{bdoc}
\alias{bdoc}
\title{Create documentation for a function}
\usage{
bdoc()
}
\value{
charstring, also printed via \code{\link{message}}
}
\description{
Create documentation for a function
}
\seealso{
\code{\link{selectObject}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr 2019
}
