% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bknit2pdf.R
\name{bknit2pdf}
\alias{bknit2pdf}
\title{pass Rnw document to knit2pdf}
\usage{
bknit2pdf(open = TRUE, clean = TRUE, ...)
}
\arguments{
\item{open}{Open pdf file?}

\item{clean}{Clean up temporary files? DEFAULT: TRUE (unlike in tools::texi2pdf)}

\item{\dots}{Further arguments passed to \code{knitr::knit}}
}
\value{
invisible charstring with file name
}
\description{
pass the filename (selected window) to \code{knitPres::rnw2pdf}, see \url{https://github.com/brry/knitPres/}.
Developed when MikTex and Rstudio stopped talking to each other. 
Can be schorcut to e.g. CTRL + K
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2021
}
