% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSource.R
\name{funSource}
\alias{funSource}
\title{Source code of a function}
\usage{
funSource(x = selectobject()$fullcode, character.only = is.character(x),
  trydirect = TRUE)
}
\arguments{
\item{x}{function name, with or without quotation marks.
DEFAULT from \code{\link{selectobject}}}

\item{character.only}{If TRUE, look for SomeFun instead of MyFun if
MyFun <- "SomeFun". DEFAULT: \code{\link{is.character}(x)}}

\item{trydirect}{If TRUE, try direct url to file \code{x.R}. DEFAULT: TRUE}
}
\value{
links that are also opened with \code{\link{browseURL}}
}
\description{
open source code of a function in a loaded or specified package
         on github.com/cran or github.com/wch/r-source
         It can be very useful to bind this addin to F7, for example.
}
\examples{
# Go to Addins - browse Addins - Keyboard shortcuts - map "funSource of an object" to F7

}
\seealso{
\code{\link{selectobject}, \link{str_addin}, \link{head_addin}, \link{View_addin}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan+Dec 2016, May 2017
}
