% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_addin.R
\name{head_addin}
\alias{head_addin}
\title{Rstudio head on F$}
\usage{
head_addin(obj = selectobject())
}
\arguments{
\item{obj}{Some R object. DEFAULT: Rstudio addin selected code}
}
\value{
Output from \code{\link{head}}.
}
\description{
Rstudio addin to show the \code{\link{head}} of a highlighted object.
Instead of an object name, code creating data can be highlighted as well.
It can be very useful to bind this addin to F$, for example.
}
\examples{
# Go to Addins - browse Addins - Keyboard shortcuts - map "structure of an object" to F4

}
\references{
Heavily borrowed from \url{https://github.com/digital-dharma/RStudioAddIns}
}
\seealso{
\code{\link{selectobject}}, \code{\link{str_addin}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2017
}
