% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/href.R
\name{href}
\alias{href}
\title{wrap URL in Rnw source code in "\\href{link}{text}"}
\usage{
href()
}
\description{
wrap URL in Rnw source code in "\\href{link}{text}"
}
\seealso{
\code{\link{selectobject}}, \code{\link{rcode}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
