% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pycheck.R
\name{pycheck}
\alias{pycheck}
\title{run python exercise development test}
\usage{
pycheck()
}
\description{
run python exercise development test
}
\seealso{
\code{\link{selectobject}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2023
}
