% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode.R
\name{rcode}
\alias{rcode}
\title{wrap Rnw source code in "\\rcode{}"}
\usage{
rcode()
}
\description{
wrap Rnw source code in "\\rcode{}"
}
\seealso{
\code{\link{selectobject}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2019
}
