% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score exercises in the course Fundamentals of Programming}
\usage{
score()
}
\description{
score exercises in the course Fundamentals of Programming
}
\seealso{
\code{\link{selectobject}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2023 / Mar 2026
}
