% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectobject.R
\name{selectobject}
\alias{selectobject}
\title{Select object}
\usage{
selectobject()
}
\value{
List with the (evaluated) object and the code generating it as a character string
}
\description{
Rstudio addin to create object from highlighted object name or code
}
\examples{
# see str_addin


}
\references{
Heavily borrowed from \url{https://github.com/digital-dharma/RStudioAddIns}
}
\seealso{
str_addin
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2017
}
