% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKeyboardBindings.R
\name{setKeyboardBindings}
\alias{setKeyboardBindings}
\title{Set Rstudio keyboard bindings}
\usage{
setKeyboardBindings(openfolder = TRUE)
}
\arguments{
\item{openfolder}{Logical: Open folders after writing the files?
Uses \code{berryFunctions\link[berryFunctions]{openFile}()}. DEFAULT: TRUE}
}
\value{
Returns nothing
}
\description{
Set Rstudio keyboard bindings as mapped 
             on \url{https://github.com/brry/rskey#rskey}.
}
\seealso{
\code{\link{addins}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2019, May 2026
}
