% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umlaut2ascii.R
\name{umlaut2ascii}
\alias{umlaut2ascii}
\title{Umlaut to ASCII}
\usage{
umlaut2ascii()
}
\value{
path, invisibly
}
\description{
Convert Umlaute in Script to ASCII. Less aggressive than 
<https://github.com/dreamRs/prefixer/#not-ascii>
}
\seealso{
\code{berryFunctions\link[berryFunctions]{convertUmlaut}()}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2022
}
\keyword{file}
