% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_addin.R
\name{View_addin}
\alias{View_addin}
\title{Rstudio View on F6}
\usage{
View_addin(obj = selectobject())
}
\arguments{
\item{obj}{Some R object. DEFAULT: Rstudio addin selected code}
}
\value{
Invisible Null, opens window as per \code{\link{View}}.
}
\description{
Rstudio addin to \code{\link{View}} a highlighted object.
Instead of an object name, code creating data can be highlighted as well.
It can be very useful to bind this addin to F6, for example.
}
\examples{
# Go to Addins - browse Addins - Keyboard shortcuts - map "View an object" to F6

}
\references{
Heavily borrowed from \url{https://github.com/digital-dharma/RStudioAddIns}
}
\seealso{
\code{\link{selectobject}}, \code{\link{str_addin}}, \code{\link{head_addin}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2017
}
