% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdoc.R
\name{bdoc}
\alias{bdoc}
\title{Create documentation for a function}
\usage{
bdoc()
}
\value{
invisible charstring, but mainly used for writing into the source file directly
}
\description{
Create Roxygen documentation skeleton for a function
}
\seealso{
\code{\link{selectobject}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr 2019
}
