% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectobject.R
\name{selectobject}
\alias{selectobject}
\title{Select object}
\usage{
selectobject(eval = TRUE)
}
\arguments{
\item{eval}{Should text be evaluated? 
If FALSE, the output is a charstring of the marked text.
DEFAULT: TRUE}
}
\value{
List with the (evaluated) object and the code generating it as a character string
}
\description{
Rstudio addin to create object from highlighted object name or code.
Also works in browser mode, albeit not in the code_browser window.
}
\examples{
# see str_addin

}
\references{
Heavily borrowed from digital-dharma/RStudioAddIns (no longer online)
}
\seealso{
\code{\link{addins}, \link{funSource}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2017
}
